Move k8s utils into new package
