+++ pkg/new/kubetest/kubeapi.go
@@ -1,7 +1,7 @@
 package kubetest
 
 import (
-	"github.com/onosproject/onos-test/pkg/util/k8s"
+	"github.com/onosproject/onos-test/pkg/new/util/k8s"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
