+++ pkg/new/kubetest/worker.go
@@ -16,7 +16,7 @@ package kubetest
 
 import (
 	"fmt"
-	"github.com/onosproject/onos-test/pkg/util/k8s"
+	"github.com/onosproject/onos-test/pkg/new/util/k8s"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"testing"
 )
